[FEAT/#291] SpoonLocalRepository 삭제
