+++ app/src/main/java/com/spoony/spoony/data/datasource/local/SpoonLocalDataSource.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface SpoonLocalDataSource {
     suspend fun getLastEntryDate(): Flow<String?>
-    suspend fun getIsSpoonDrawn(): Flow<Boolean?>
+    suspend fun getIsSpoonDrawn(): Flow<Boolean>
     suspend fun updateLastEntryDate(date: String)
     suspend fun updateSpoonDrawn(isSpoonDrawn: Boolean)
 }
