--- app/src/main/java/com/spoony/spoony/domain/repository/SpoonLocalRepository.kt
@@ -1,7 +1,0 @@
-package com.spoony.spoony.domain.repository
-
-interface SpoonLocalRepository {
-    suspend fun getSpoonDrawLog(): Pair<String?, Boolean?>
-    suspend fun updateLastEntryDate(date: String)
-    suspend fun updateSpoonDrawn()
-}
