+++ app/src/main/java/com/spoony/spoony/domain/repository/SpoonRepository.kt
@@ -7,4 +7,8 @@ interface SpoonRepository {
     suspend fun drawSpoon(): Result<SpoonEntity>
     suspend fun getWeeklySpoonDraw(): Result<SpoonListEntity>
     suspend fun getSpoonCount(): Result<Int>
+
+    suspend fun getSpoonDrawLog(): Pair<String?, Boolean>
+    suspend fun updateLastEntryDate(date: String)
+    suspend fun updateSpoonDrawn()
 }
